feat: protect collaboration actions
only owner can enable/disable/remove a collaboration request

+++ packages/api/src/collaboration/collaboration.service.ts
@@ -6,6 +6,9 @@ import { User } from '../users/user.entity';
 import { Collaboration } from './collaboration.model';
 import { UserRepository } from '../users/user.repository';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Document } from '../document/document.model';
+import { DocumentService } from '../document/document.service';
+import { UnauthorizedException } from '@nestjs/common/exceptions/unauthorized.exception';
 
 @Injectable()
 export class CollaborationService {
@@ -14,26 +17,34 @@ export class CollaborationService {
     private readonly collaborationModel: Model<Collaboration>,
     @InjectRepository(UserRepository)
     private userRepository?: UserRepository,
+    @InjectModel('Document') private readonly documentModel?: Model<Document>,
   ) {}
 
   async collaborationTeam(documentId: string) {
-    const collaborations = await this.collaborationModel.find({documentId, state:true});
-    const ids:string[]=collaborations.map((collaboration)=>collaboration.userId)
+    const collaborations = await this.collaborationModel.find({
+      documentId,
+      state: true,
+    });
+    const ids: string[] = collaborations.map(
+      collaboration => collaboration.userId,
+    );
     return await this.userRepository.findByIds(ids);
-
   }
 
   async collaborationRequests(documentId: string) {
-    const collaborations = await this.collaborationModel.find({documentId, state:false});
-/*     const result= { [documentId]: collaborations.map(async (c)=>await this.userRepository.findOne(c.userId)) }
+    const collaborations = await this.collaborationModel.find({
+      documentId,
+      state: false,
+    });
+    /*     const result= { [documentId]: collaborations.map(async (c)=>await this.userRepository.findOne(c.userId)) }
     return result */
-    const ids:string[]=collaborations.map((collaboration)=>collaboration.userId)
-    return {[documentId]: await this.userRepository.findByIds(ids)}
-
+    const ids: string[] = collaborations.map(
+      collaboration => collaboration.userId,
+    );
+    return { [documentId]: await this.userRepository.findByIds(ids) };
   }
 
-
-  async joinTeam(currentUser: User, id: string, isOwner:boolean) {
+  async joinTeam(currentUser: User, id: string, isOwner: boolean) {
     const collaboration = new this.collaborationModel();
     collaboration.userId = currentUser.id;
     collaboration.documentId = id;
@@ -56,38 +67,66 @@ export class CollaborationService {
 
   async isMember(currentUser: User, id: string) {
     return await this.collaborationModel
-      .findOne({ userId: currentUser.id, documentId: id, state:true })
+      .findOne({ userId: currentUser.id, documentId: id, state: true })
       .exec();
   }
 
-  async teamCount(id:string)
-  {
-    return await this.collaborationModel.find({documentId:id , state:true}).count();
+  async teamCount(id: string) {
+    return await this.collaborationModel
+      .find({ documentId: id, state: true })
+      .count();
   }
 
-  async enable(userId:string, docId:string)
-  {
-    const user = await this.collaborationModel.findOne({documentId:docId , userId:userId})
-    user.state = true;
-    await user.save();
-    return user;
+  async enable(currentUser: User, userId: string, docId: string) {
+    if (
+      await new DocumentService(
+        this.documentModel,
+        this.collaborationModel,
+        this.userRepository,
+      ).isOwner(currentUser, docId)
+    ) {
+      const user = await this.collaborationModel.findOne({
+        documentId: docId,
+        userId: userId,
+      });
+      user.state = true;
+      await user.save();
+      return user;
+    } else throw new UnauthorizedException('You need to be the owner');
   }
 
-  async disable(userId:string, docId:string)
-  {
-    const user = await this.collaborationModel.findOne({documentId:docId , userId:userId})
+  async disable(currentUser: User, userId: string, docId: string) {
+    if (
+      await new DocumentService(
+        this.documentModel,
+        this.collaborationModel,
+        this.userRepository,
+      ).isOwner(currentUser, docId)
+    ) {
+    const user = await this.collaborationModel.findOne({
+      documentId: docId,
+      userId: userId,
+    });
     user.state = false;
     await user.save();
     return user;
+  } else throw new UnauthorizedException('You need to be the owner');
   }
 
-  async remove(userId:string, docId:string)
-  {
+  async remove(currentUser: User, userId: string, docId: string) {
+    if (
+      await new DocumentService(
+        this.documentModel,
+        this.collaborationModel,
+        this.userRepository,
+      ).isOwner(currentUser, docId)
+    ) {
     const del = await this.collaborationModel
-    .deleteOne({ userId: userId, documentId: docId })
-    .exec();
-  if (del.n === 0) {
-    throw new NotFoundException();
-  }
+      .deleteOne({ userId: userId, documentId: docId })
+      .exec();
+    if (del.n === 0) {
+      throw new NotFoundException();
+    }
+  } else throw new UnauthorizedException('You need to be the owner');
   }
 }
